requirements.txt
numpy 
random
matplotlib.pyplot
seaborn